// play/chess-tourney-frontend/src/api/model/updateProfilePayload.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */

export interface UpdateProfilePayload {
  /**
   * @minimum 100
   * @maximum 3000
   */
  lichess_rating?: number;
  /**
   * @minimum 100
   * @maximum 3000
   */
  chess_com_rating?: number;
  /**
   * @minimum 100
   * @maximum 3000
   */
  fide_rating?: number;
  /**
   * @minimum 0
   * @maximum 10
   */
  experience_with_chess?: number;
}